slight refactoring of tests
