+++ test/common.hpp
@@ -2,6 +2,7 @@
 
 #include <memory>
 
+#include "zombie/zombie.hpp"
 #include "zombie/heap/heap.hpp"
 
 template<bool is_unique>
@@ -44,4 +45,44 @@ template<bool is_unique>
 struct NotifyIndexChanged {
   void operator()(const Element<is_unique>&, const size_t&) {
   }
+};
+
+
+
+
+
+template<>
+struct GetSize<int> {
+  size_t operator()(const int&) {
+    return sizeof(int);
+  }
+};
+
+
+
+// [test_id] is used to separate different tests
+template<typename test_id>
+struct Resource {
+  static unsigned int count;
+  static unsigned int destructor_count;
+
+  int value;
+
+  Resource(int value) : value(value) { ++count; }
+  Resource(const Resource&) = delete;
+  ~Resource() { --count; ++destructor_count; }
+};
+
+template<typename test_id>
+unsigned int Resource<test_id>::count = 0;
+
+template<typename test_id>
+unsigned int Resource<test_id>::destructor_count = 0;
+
+
+template<typename test_id>
+struct GetSize<Resource<test_id>> {
+  size_t operator()(const Resource<test_id>&) {
+    return 1 << 16;
+  };
 };
